ex15 modificado

+++ PHP-ARRAYS/php-b1/ex-06.php
@@ -6,7 +6,7 @@
 
     for ($i = 0; $i < 9; $i++){
         echo "Escreva o " . $i + 2 . "º número: ";
-        $num = trim(fgets(STDIN));
+        $num = (int)trim(fgets(STDIN));
         if ($num < $menorNum){
             $menorNum = $num;
         }
